copied online filtering from preprocessing to lib/EEGsynth
